remate bigint to int

+++ src/app/models/rental_requests/rental_history.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import date, datetime
 
-from sqlalchemy import BigInteger, Date, DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import Date, DateTime, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database.base import Base
@@ -13,7 +13,7 @@ class RentalHistory(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     account_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("accounts.id", name="fk_rental_history_account"),
         nullable=False,
     )
